Backoff module extraction
