+++ main/src/main/java/io/split/android/client/service/sseclient/sseclient/BackoffCounterTimer.java
@@ -8,7 +8,7 @@ import io.split.android.client.service.executor.SplitTask;
 import io.split.android.client.service.executor.SplitTaskExecutionInfo;
 import io.split.android.client.service.executor.SplitTaskExecutionListener;
 import io.split.android.client.service.executor.SplitTaskExecutor;
-import io.split.android.client.service.sseclient.BackoffCounter;
+import io.split.android.client.backoff.BackoffCounter;
 import io.split.android.client.utils.logger.Logger;
 
 public class BackoffCounterTimer implements SplitTaskExecutionListener {
