+++ main/src/main/java/io/split/android/client/service/sseclient/sseclient/RetryBackoffCounterTimer.java
@@ -11,7 +11,7 @@ import io.split.android.client.service.executor.SplitTaskExecutionInfo;
 import io.split.android.client.service.executor.SplitTaskExecutionListener;
 import io.split.android.client.service.executor.SplitTaskExecutionStatus;
 import io.split.android.client.service.executor.SplitTaskExecutor;
-import io.split.android.client.service.sseclient.BackoffCounter;
+import io.split.android.client.backoff.BackoffCounter;
 import io.split.android.client.utils.logger.Logger;
 
 import java.util.concurrent.TimeUnit;
